pyproject.toml
[project]
name = "async-alchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "greenlet>=3.1.1",
    "pydantic>=2.10.5",
    "sqlalchemy>=2.0.37",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = ["app"]
